Hangman Game Project
The hamgman.py file imports all the words in the wordlist.py file and a random word from this list of words is assigned to a variable as long as the word meets a certain criteria described in the function.

A random word is assigned each time the loop runs and the player is asked to guess the word by entering letters and those letters are compared to the set of letters that make up the right answer. If the letters entered form the right word before there are no more lives, the player wins!